[30856] Reposition ng-select
https://community.openproject.com/wp/30856
