+++ frontend/src/app/components/wp-fast-table/builders/internal-sort-columns.ts
@@ -1,0 +1,10 @@
+import {QueryColumn} from "core-components/wp-query/query-column";
+
+export const internalSortColumn = {
+  id: '__internal-sorthandle'
+} as QueryColumn;
+
+export const internalContextMenuColumn = {
+  id: '__internal-contextMenu'
+} as QueryColumn;
+
